figures created

+++ Python/Programs/data_import.py
@@ -124,11 +124,12 @@ heathrow_code = pd.DataFrame(data={'Code' : ['EHEATHROW'],
                                    'Population' : ['206,800']})
 population_statistics_uncleaned_df = pd.concat([population_statistics_uncleaned_df, heathrow_code], axis=0, ignore_index=True)
 
+
 ## Structure data types and indices for loaded data appropriately
 
 # Correct default loaded data types
 #- Transform population comma-separated string data type from string to int: first regex to replace non-integers with nothing, then convert to numeric
-population_statistics_uncleaned_df['Population'] = pd.to_numeric(population_statistics_uncleaned_df['Population'].str.replace('[^0-9]', ''))
+population_statistics_uncleaned_df['Population'] = pd.to_numeric(population_statistics_uncleaned_df['Population'].str.replace('[^0-9]', '')).astype('int')
 
 # Make Code column the index - tested and contains no duplicates so safe to do so
 #- 0 Duplicates in this index, makes sense as one area should have one code - can use without worry
